removed old file based classes and then replaced with in Memory versions and renamed in memory versions
