+++ tests/test_cs_transform.py
@@ -29,7 +29,7 @@ class TestCSTransformService(unittest.TestCase):
 
         return z.namelist()
 
-    @patch('transform.transformers.InMemoryImageTransformer.get_image_sequence_list', return_value=[13, 14])
+    @patch('transform.transformers.ImageTransformer.get_image_sequence_list', return_value=[13, 14])
     def test_creates_cs_defaults(self, mock_sequence_no):
 
         ziplist = self.get_zip_list(self.transform_cs_endpoint)
@@ -45,7 +45,7 @@ class TestCSTransformService(unittest.TestCase):
 
         self.assertEqual(expected, ziplist)
 
-    @patch('transform.transformers.InMemoryImageTransformer.get_image_sequence_list', return_value=[1985, 1986])
+    @patch('transform.transformers.ImageTransformer.get_image_sequence_list', return_value=[1985, 1986])
     def test_creates_cs_sequence(self, mock_sequence_no):
 
         ziplist = self.get_zip_list(self.transform_cs_endpoint + "/2345")
