+++ transform/transformers/__init__.py
@@ -1,11 +1,9 @@
-from .InMemoryCSTransformer import InMemoryCSTransformer
+from .CSTransformer import CSTransformer
 from .PDFTransformer import PDFTransformer
 from .ImageTransformer import ImageTransformer
-from .InMemoryImageTransformer import InMemoryImageTransformer
 from .PCKTransformer import PCKTransformer
 from .MWSSTransformer import MWSSTransformer
-from .InMemoryMWSSTransformer import InMemoryMWSSTransformer
 
 __all__ = [
-    'PDFTransformer', 'ImageTransformer', 'InMemoryImageTransformer', 'PCKTransformer',
-    'MWSSTransformer', 'InMemoryCSTransformer', 'InMemoryMWSSTransformer']
+    'PDFTransformer', 'ImageTransformer', 'PCKTransformer',
+    'CSTransformer', 'MWSSTransformer']
